adding spring security and ownership shecks

+++ src/main/java/com/example/resumlik/service/AddressService.java
@@ -3,8 +3,10 @@ package com.example.resumlik.service;
 import com.example.resumlik.dto.request.AddressRequestDto;
 import com.example.resumlik.dto.response.AddressResponseDto;
 import com.example.resumlik.model.Address;
+import com.example.resumlik.model.Resume;
 import com.example.resumlik.repository.AddressRepository;
 import com.example.resumlik.repository.ResumeRepository;
+import com.example.resumlik.util.AuthHelper;
 import com.example.resumlik.util.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -33,9 +35,13 @@ public class AddressService {
     }
 
     public Response<AddressResponseDto> create(AddressRequestDto addressRequestDto) {
+        Resume resume = resumeRepository.findById(addressRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found"));
+        if (!AuthHelper.checkOwnerShip(resume)){
+            throw new RuntimeException("You are not authorized to create address for this resume");
+        }
         Address address = addressRequestDto.toEntity();
 
-        address.setResume(resumeRepository.findById(addressRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found")));
+        address.setResume(resume);
 
         return Response.<AddressResponseDto>builder()
                 .result(new AddressResponseDto(addressRepository.save(address)))
@@ -45,7 +51,11 @@ public class AddressService {
 
 
     public Response<String> delete(Long id) {
-        addressRepository.findById(id).orElseThrow(() -> new RuntimeException("Address not found"));
+        Address address = addressRepository.findById(id).orElseThrow(() -> new RuntimeException("Address not found"));
+
+        if (!AuthHelper.checkOwnerShip(address)){
+            throw new RuntimeException("You are not authorized to delete this address");
+        }
 
         addressRepository.deleteById(id);
 
@@ -56,6 +66,10 @@ public class AddressService {
 
     public Response<AddressResponseDto> update(Long id, AddressRequestDto addressRequestDto) {
         Address address = addressRepository.findById(id).orElseThrow(() -> new RuntimeException("Address not found"));
+
+        if (!AuthHelper.checkOwnerShip(address)){
+            throw new RuntimeException("You are not authorized to update this address");
+        }
 
         address.setStreet(addressRequestDto.getStreet());
         address.setAdditionalStreet(addressRequestDto.getAdditionalStreet());
