+++ src/main/java/com/example/resumlik/service/CertificateService.java
@@ -3,10 +3,12 @@ package com.example.resumlik.service;
 import com.example.resumlik.dto.request.CertificateRequestDto;
 import com.example.resumlik.dto.response.CertificateResponseDto;
 import com.example.resumlik.model.Certificate;
+import com.example.resumlik.model.Resume;
 import com.example.resumlik.model.Skill;
 import com.example.resumlik.repository.CertificateRepository;
 import com.example.resumlik.repository.ResumeRepository;
 import com.example.resumlik.repository.SkillRepository;
+import com.example.resumlik.util.AuthHelper;
 import com.example.resumlik.util.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,48 +18,17 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class CertificateService {
-//    public final SkillRepository skillRepository;
-//    private final ResumeRepository resumeRepository;
-//
-//    public Response<List<SkillResponseDto>> getAllByResumeId(Long resumeId) {
-//
-//        List<SkillResponseDto> skillResponseDtos = skillRepository.findAllByResumeId(resumeId)
-//                .stream()
-//                .map(SkillResponseDto::new)
-//                .toList();
-//
-//        return Response.<List<SkillResponseDto>>builder()
-//                .result(skillResponseDtos)
-//                .build();
-//    }
-//
-//    public Response<SkillResponseDto> save(SkillRequestDto skillRequestDto) {
-//
-//        Skill skill = skillRequestDto.toEntity();
-//        skill.setResume(resumeRepository.findById(skillRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found.")));
-//
-//        return Response.<SkillResponseDto>builder()
-//                .result(new SkillResponseDto(skillRepository.save(skill)))
-//                .build();
-//    }
-//
-//    public Response<String> delete(Long id) {
-//        skillRepository.findById(id).orElseThrow(() -> new RuntimeException("Skill not found."));
-//
-//        skillRepository.deleteById(id);
-//
-//        return Response.<String>builder()
-//                .message("Skill deleted.")
-//                .result("Skill deleted.")
-//                .build();
-//    } follow this pattern to create a new service
 
     private final CertificateRepository certificateRepository;
     private final ResumeRepository resumeRepository;
 
     public Response<CertificateResponseDto> save(CertificateRequestDto certificateRequestDto) {
+        Resume resume = resumeRepository.findById(certificateRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found."));
+        if (!AuthHelper.checkOwnerShip(resume)){
+            throw new RuntimeException("You are not authorized to create certificate for this resume.");
+        }
         Certificate certificate = certificateRequestDto.toEntity();
-        certificate.setResume(resumeRepository.findById(certificateRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found.")));
+        certificate.setResume(resume);
 
         return Response.<CertificateResponseDto>builder()
                 .result(
@@ -68,7 +39,12 @@ public class CertificateService {
     }
 
     public Response<String> delete(Long id) {
-        certificateRepository.findById(id).orElseThrow(() -> new RuntimeException("Certificate not found."));
+        Certificate certificate = certificateRepository.findById(id).orElseThrow(() -> new RuntimeException("Certificate not found."));
+
+        if (!AuthHelper.checkOwnerShip(certificate)){
+            throw new RuntimeException("You are not authorized to delete this certificate.");
+        }
+
 
         certificateRepository.deleteById(id);
 
@@ -89,4 +65,19 @@ public class CertificateService {
                 .build();
     }
 
+    public Response<CertificateResponseDto> update(Long id, CertificateRequestDto certificateRequestDto) {
+        Certificate certificate = certificateRepository.findById(id).orElseThrow(() -> new RuntimeException("Certificate not found."));
+        if (!AuthHelper.checkOwnerShip(certificate)){
+            throw new RuntimeException("You are not authorized to update this certificate.");
+        }
+        certificate.setName(certificateRequestDto.getName());
+        certificate.setIssuer(certificateRequestDto.getIssuer());
+        certificate.setDate(certificateRequestDto.getDate());
+
+        return Response.<CertificateResponseDto>builder()
+                .result(new CertificateResponseDto(certificateRepository.save(certificate)))
+                .message("Certificate updated.")
+                .build();
+
+    }
 }
