+++ src/main/java/com/example/resumlik/service/ExperienceService.java
@@ -3,8 +3,10 @@ package com.example.resumlik.service;
 import com.example.resumlik.dto.request.ExperienceRequestDto;
 import com.example.resumlik.dto.response.ExperienceResponseDto;
 import com.example.resumlik.model.Experience;
+import com.example.resumlik.model.Resume;
 import com.example.resumlik.repository.ExperienceRepository;
 import com.example.resumlik.repository.ResumeRepository;
+import com.example.resumlik.util.AuthHelper;
 import com.example.resumlik.util.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,42 +16,6 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class ExperienceService {
-
-//    public final SkillRepository skillRepository;
-//    private final ResumeRepository resumeRepository;
-//
-//    public Response<List<SkillResponseDto>> getAllByResumeId(Long resumeId) {
-//
-//        List<SkillResponseDto> skillResponseDtos = skillRepository.findAllByResumeId(resumeId)
-//                .stream()
-//                .map(SkillResponseDto::new)
-//                .toList();
-//
-//        return Response.<List<SkillResponseDto>>builder()
-//                .result(skillResponseDtos)
-//                .build();
-//    }
-//
-//    public Response<SkillResponseDto> save(SkillRequestDto skillRequestDto) {
-//
-//        Skill skill = skillRequestDto.toEntity();
-//        skill.setResume(resumeRepository.findById(skillRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found.")));
-//
-//        return Response.<SkillResponseDto>builder()
-//                .result(new SkillResponseDto(skillRepository.save(skill)))
-//                .build();
-//    }
-//
-//    public Response<String> delete(Long id) {
-//        skillRepository.findById(id).orElseThrow(() -> new RuntimeException("Skill not found."));
-//
-//        skillRepository.deleteById(id);
-//
-//        return Response.<String>builder()
-//                .message("Skill deleted.")
-//                .result("Skill deleted.")
-//                .build();
-//    }
 
     public final ExperienceRepository experienceRepository;
     private final ResumeRepository resumeRepository;
@@ -67,9 +33,12 @@ public class ExperienceService {
     }
 
     public Response<ExperienceResponseDto> save(ExperienceRequestDto experienceRequestDto) {
-
+        Resume resume = resumeRepository.findById(experienceRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found."));
+        if(!AuthHelper.checkOwnerShip(resume)){
+            throw new RuntimeException("You are not authorized to create experience for this resume.");
+        }
         Experience experience = experienceRequestDto.toEntity();
-        experience.setResume(resumeRepository.findById(experienceRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found.")));
+        experience.setResume(resume);
 
         return Response.<ExperienceResponseDto>builder()
                 .result(new ExperienceResponseDto(experienceRepository.save(experience)))
@@ -77,7 +46,10 @@ public class ExperienceService {
     }
 
     public Response<String> delete(Long id) {
-        experienceRepository.findById(id).orElseThrow(() -> new RuntimeException("Experience not found."));
+        Experience experience = experienceRepository.findById(id).orElseThrow(() -> new RuntimeException("Experience not found."));
+        if(!AuthHelper.checkOwnerShip(experience)){
+            throw new RuntimeException("You are not authorized to delete this experience.");
+        }
 
         experienceRepository.deleteById(id);
 
@@ -89,6 +61,9 @@ public class ExperienceService {
 
     public Response<ExperienceResponseDto> update(Long id, ExperienceRequestDto experienceRequestDto) {
         Experience experience = experienceRepository.findById(id).orElseThrow(() -> new RuntimeException("Experience not found."));
+        if(!AuthHelper.checkOwnerShip(experience)){
+            throw new RuntimeException("You are not authorized to update this experience.");
+        }
 
         experience.setCompany(experienceRequestDto.getCompany());
         experience.setTitle(experienceRequestDto.getTitle());
