+++ src/main/java/com/example/resumlik/service/HobbyService.java
@@ -3,8 +3,10 @@ package com.example.resumlik.service;
 import com.example.resumlik.dto.request.HobbyRequestDto;
 import com.example.resumlik.dto.response.HobbyResponseDto;
 import com.example.resumlik.model.Hobby;
+import com.example.resumlik.model.Resume;
 import com.example.resumlik.repository.HobbyRepository;
 import com.example.resumlik.repository.ResumeRepository;
+import com.example.resumlik.util.AuthHelper;
 import com.example.resumlik.util.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,8 +30,12 @@ public class HobbyService {
     }
 
     public Response<HobbyResponseDto> save(HobbyRequestDto hobbyRequestDto) {
+        Resume resume = resumeRepository.findById(hobbyRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found."));
+        if (!AuthHelper.checkOwnerShip(resume)){
+            throw new RuntimeException("You are not authorized to create hobby for this resume.");
+        }
         Hobby hobby = hobbyRequestDto.toEntity();
-        hobby.setResume(resumeRepository.findById(hobbyRequestDto.getResumeId()).orElseThrow(()->new RuntimeException("Resume not found")));
+        hobby.setResume(resume);
 
         return Response.<HobbyResponseDto>builder()
                 .result(new HobbyResponseDto(hobbyRepository.save(hobby)))
@@ -38,12 +44,28 @@ public class HobbyService {
     }
 
     public Response<String> delete(Long id) {
-        hobbyRepository.findById(id).orElseThrow(()->new RuntimeException("Hobby not found"));
+        Hobby hobby = hobbyRepository.findById(id).orElseThrow(()->new RuntimeException("Hobby not found"));
+        if (!AuthHelper.checkOwnerShip(hobby)){
+            throw new RuntimeException("You are not authorized to delete this hobby.");
+        }
 
         hobbyRepository.deleteById(id);
 
         return Response.<String>builder()
                 .message("Hobby deleted successfully")
+                .build();
+    }
+
+    public Response<HobbyResponseDto> update(Long id, HobbyRequestDto hobbyRequestDto) {
+        Hobby hobby = hobbyRepository.findById(id).orElseThrow(()->new RuntimeException("Hobby not found"));
+        if (!AuthHelper.checkOwnerShip(hobby)){
+            throw new RuntimeException("You are not authorized to update this hobby.");
+        }
+        hobby.setName(hobbyRequestDto.getName());
+
+        return Response.<HobbyResponseDto>builder()
+                .result(new HobbyResponseDto(hobbyRepository.save(hobby)))
+                .message("Hobby updated successfully")
                 .build();
     }
 }
