+++ src/main/java/com/example/resumlik/service/LanguageService.java
@@ -3,8 +3,10 @@ package com.example.resumlik.service;
 import com.example.resumlik.dto.request.LanguageRequestDto;
 import com.example.resumlik.dto.response.LanguageResponseDto;
 import com.example.resumlik.model.Language;
+import com.example.resumlik.model.Resume;
 import com.example.resumlik.repository.LanguageRepository;
 import com.example.resumlik.repository.ResumeRepository;
+import com.example.resumlik.util.AuthHelper;
 import com.example.resumlik.util.Response;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -33,9 +35,12 @@ public class LanguageService {
     }
 
     public Response<LanguageResponseDto> save(LanguageRequestDto languageRequestDto) {
-
+        Resume resume = resumeRepository.findById(languageRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found."));
+        if (!AuthHelper.checkOwnerShip(resume)){
+            throw new RuntimeException("You are not authorized to create language for this resume.");
+        }
         Language language = languageRequestDto.toEntity();
-        language.setResume(resumeRepository.findById(languageRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found.")));
+        language.setResume(resume);
 
         return Response.<LanguageResponseDto>builder()
                 .result(new LanguageResponseDto(languageRepository.save(language)))
@@ -43,11 +48,27 @@ public class LanguageService {
     }
 
     public Response<String> delete(Long id) {
-        languageRepository.findById(id).orElseThrow(() -> new RuntimeException("Language not found."));
+        Language language = languageRepository.findById(id).orElseThrow(() -> new RuntimeException("Language not found."));
+        if (!AuthHelper.checkOwnerShip(language)){
+            throw new RuntimeException("You are not authorized to delete this language.");
+        }
 
         languageRepository.deleteById(id);
         return Response.<String>builder()
                 .result("Language deleted.")
+                .build();
+    }
+
+    public Response<LanguageResponseDto> update(Long id, @Valid LanguageRequestDto languageRequestDto) {
+        Language language = languageRepository.findById(id).orElseThrow(() -> new RuntimeException("Language not found."));
+        if (!AuthHelper.checkOwnerShip(language)){
+            throw new RuntimeException("You are not authorized to update this language.");
+        }
+        language.setName(languageRequestDto.getName());
+        language.setLevel(languageRequestDto.getLevel());
+
+        return Response.<LanguageResponseDto>builder()
+                .result(new LanguageResponseDto(languageRepository.save(language)))
                 .build();
     }
 }
