+++ src/main/java/com/example/resumlik/service/ProjectService.java
@@ -3,9 +3,12 @@ package com.example.resumlik.service;
 
 import com.example.resumlik.dto.request.ProjectRequestDto;
 import com.example.resumlik.dto.response.ProjectResponseDto;
+import com.example.resumlik.enums.ProjectMode;
 import com.example.resumlik.model.Project;
+import com.example.resumlik.model.Resume;
 import com.example.resumlik.repository.ProjectRepository;
 import com.example.resumlik.repository.ResumeRepository;
+import com.example.resumlik.util.AuthHelper;
 import com.example.resumlik.util.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -55,9 +58,12 @@ public class ProjectService {
     private final ResumeRepository resumeRepository;
 
     public Response<ProjectResponseDto> save(ProjectRequestDto projectRequestDto) {
-
+        Resume resume = resumeRepository.findById(projectRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found."));
+        if (!AuthHelper.checkOwnerShip(resume)){
+            throw new RuntimeException("You are not authorized to create project for this resume.");
+        }
         Project project = projectRequestDto.toEntity();
-        project.setResume(resumeRepository.findById(projectRequestDto.getResumeId()).orElseThrow(() -> new RuntimeException("Resume not found.")));
+        project.setResume(resume);
 
         return Response.<ProjectResponseDto>builder()
                 .result(new ProjectResponseDto(projectRepository.save(project)))
@@ -77,13 +83,33 @@ public class ProjectService {
     }
 
     public Response<String> delete(Long id) {
-        projectRepository.findById(id).orElseThrow(() -> new RuntimeException("Project not found."));
+        Project project =  projectRepository.findById(id).orElseThrow(() -> new RuntimeException("Project not found."));
+        if (!AuthHelper.checkOwnerShip(project)){
+            throw new RuntimeException("You are not authorized to delete this project.");
+        }
 
         projectRepository.deleteById(id);
 
         return Response.<String>builder()
                 .message("Project deleted.")
                 .result("Project deleted.")
+                .build();
+    }
+
+    public Response<ProjectResponseDto> update(Long id, ProjectRequestDto projectRequestDto) {
+        Project project = projectRepository.findById(id).orElseThrow(() -> new RuntimeException("Project not found."));
+        if (!AuthHelper.checkOwnerShip(project)){
+            throw new RuntimeException("You are not authorized to update this project.");
+        }
+        project.setName(projectRequestDto.getName());
+        project.setMode(ProjectMode.valueOf(projectRequestDto.getMode()));
+        project.setRealisedAt(projectRequestDto.getRealisedAt());
+        project.setDescription(projectRequestDto.getDescription());
+        project.setTitle(projectRequestDto.getTitle());
+
+        return Response.<ProjectResponseDto>builder()
+                .result(new ProjectResponseDto(projectRepository.save(project)))
+                .message("Project updated.")
                 .build();
     }
 }
